+++ workspaces/tekton/plugins/tekton/CHANGELOG.md
@@ -30,6 +30,12 @@
   - @backstage-community/plugin-tekton-common@1.17.0
   - @backstage-community/plugin-tekton-react@0.2.0
 
+## 3.33.3
+
+### Patch Changes
+
+- 782b46b: Added and updated fr, it and ja translation files
+
 ## 3.33.2
 
 ### Patch Changes